+++ Changelog.md
@@ -1,7 +1,18 @@
-### 1.2.8.2
+### 1.2.9
 
 - Module *NgxExport.Tools.Resolve*.
-  + Allow setting weights for names collected in DNS *A* queries.
+  + Correctly print IP addresses in the server data collected in *A* queries.
+  + Allow setting weights to names in *QueryA* data.
+  + Allow priority policies in *QueryA* data.
+  + Now, if the number of upstreams in the priority list is greater than the
+    number of variations of weights collected in *SRV* queries or set in
+    *QueryA* data, the remainder of upstreams get servers with the lowest
+    weight rather than not being specified at all in the collected data.
+  + Now, with the *PriorityList* policy, weights of servers in the collected
+    data are deliberately not specified.
+  + Names in the *QueryA* name list may contain suffix *:port* where *port* is
+    a port number, this suffix is ignored at resolving step in *collectA* and
+    only appended to values of the collected server addresses.
 
 ### 1.2.8.1